Split the rest of the API endpoints into their own files.  Removed ApiController.  Tweaks to postman for some access control tests to work (needed to look for a specific validation error inside the message instead of assuming the whole message would be equal).
