+++ src/main/java/gov/cdc/izgateway/xform/api/SolutionApiController.java
@@ -1,0 +1,95 @@
+package gov.cdc.izgateway.xform.api;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import gov.cdc.izgateway.security.AccessControlRegistry;
+import gov.cdc.izgateway.xform.model.Solution;
+import gov.cdc.izgateway.xform.security.Roles;
+import gov.cdc.izgateway.xform.services.SolutionService;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.validation.Valid;
+import lombok.extern.java.Log;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.UUID;
+import java.util.logging.Level;
+
+@Log
+@Lazy(false)
+@RestController
+public class SolutionApiController extends BaseApiController {
+    private final SolutionService solutionService;
+
+    @Value("${xform.allow-delete-via-api}")
+    private Boolean allowDelete;
+
+    @Autowired
+    public SolutionApiController(
+            SolutionService solutionService,
+            AccessControlRegistry registry
+    ) {
+        this.solutionService = solutionService;
+        registry.register(this);
+    }
+
+    @RolesAllowed({Roles.SOLUTION_READER})
+    @GetMapping("/api/v1/solutions/{uuid}")
+    public ResponseEntity<Solution> getSolutionByUUID(@PathVariable UUID uuid) {
+        Solution entity = solutionService.getObject(uuid);
+        if (entity == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(entity, HttpStatus.OK);
+    }
+
+    @RolesAllowed({Roles.SOLUTION_WRITER})
+    @PutMapping("/api/v1/solutions/{uuid}")
+    public ResponseEntity<Solution> updateSolution(@PathVariable UUID uuid, @Valid @RequestBody Solution updatedSolution) {
+        updatedSolution.setId(uuid);
+        solutionService.update(updatedSolution);
+        return new ResponseEntity<>(updatedSolution, HttpStatus.OK);
+    }
+
+    @RolesAllowed({Roles.SOLUTION_READER})
+    @GetMapping("/api/v1/solutions")
+    public ResponseEntity<String> getSolutionsList(
+            @RequestParam(required = false) String nextCursor,
+            @RequestParam(required = false) String prevCursor,
+            @RequestParam(defaultValue = "false") Boolean includeInactive,
+            @RequestParam(defaultValue = "10") int limit
+    ) {
+        try {
+            return processList(solutionService.getList(), nextCursor, prevCursor, includeInactive, limit);
+        } catch (JsonProcessingException e) {
+            log.log(Level.SEVERE, e.getMessage(), e);
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @RolesAllowed({Roles.SOLUTION_WRITER})
+    @PostMapping("/api/v1/solutions")
+    public ResponseEntity<Solution> createSolution(
+            @Valid @RequestBody() Solution solution
+    ) {
+        solutionService.create(solution);
+        return new ResponseEntity<>(solution, HttpStatus.OK);
+    }
+
+    @RolesAllowed({Roles.SOLUTION_DELETER})
+    @DeleteMapping("/api/v1/solutions/{uuid}")
+    public ResponseEntity<Solution> deleteSolution(
+            @PathVariable UUID uuid
+    ) {
+
+        if (Boolean.FALSE.equals(allowDelete)) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+
+        solutionService.delete(uuid);
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    }
+}
